// src/generated/model/hentKravdetaljerJsonResponseKravTilleggsinformasjon.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Tilbakekreving API
 * OpenAPI spec version: 1.0
 */

export type HentKravdetaljerJsonResponseKravTilleggsinformasjon = {
  type: string;
  periode?: {
  fom: string;
  tom: string;
} | null;
  /** @nullable */
  stoppdatoForLøpendeMulkt?: string | null;
  ytelserForAvregning?: {
  valuta: string;
  beløp: number;
} | null;
  tilbakekrevingsperiode?: {
  fom: string;
  tom: string;
} | null;
} | null;